+++ lib/Common/PID_v2.h
@@ -18,34 +18,34 @@ class PID
 
   //commonly used functions **************************************************************************
     PID();
-    PID(double, double, double, int, int);   // * constructor.  links the PID to the Input, Output, and 
+    PID(float, float, float, int, int);   // * constructor.  links the PID to the Input, Output, and 
                                           //   Setpoint.  Initial tuning parameters are also set here.
                                           //   (overload for specifying proportional mode)
 
-    PID(double, double, double);           // * constructor.  links the PID to the Input, Output, and 
+    PID(float, float, float);           // * constructor.  links the PID to the Input, Output, and 
                                              //   Setpoint.  Initial tuning parameters are also set here
-    PID(double, double, double, int); 
+    PID(float, float, float, int); 
 	
     // void SetMode(int Mode);               // * sets PID to either Manual (0) or Auto (non-0)
 
-    double Compute(double currentValue, 
-                  double setPoint);        // * performs the PID calculation.  it should be
+    float Compute(float currentValue, 
+                  float setPoint);        // * performs the PID calculation.  it should be
                                           //   called every time loop() cycles. ON/OFF and
                                           //   calculation frequency can be set using SetMode
                                           //   SetSampleTime respectively
 
-    void SetOutputLimits(double, double);   // * clamps the output to a specific range. 0-255 by default, but
+    void SetOutputLimits(float, float);   // * clamps the output to a specific range. 0-255 by default, but
 										                      //   it's likely the user will want to change this depending on
 										                      //   the application
 	
 
 
   //available but not commonly used functions ********************************************************
-    void SetTunings(double, double,       // * While most users will set the tunings once in the 
-                    double);         	  //   constructor, this function gives the user the option
+    void SetTunings(float, float,       // * While most users will set the tunings once in the 
+                    float);         	  //   constructor, this function gives the user the option
                                         //   of changing tunings during runtime for Adaptive control
-    void SetTunings(double, double,       // * overload for specifying proportional mode
-                    double, int);         	  
+    void SetTunings(float, float,       // * overload for specifying proportional mode
+                    float, int);         	  
 
 	  void SetControllerDirection(int);	  // * Sets the Direction, or "Action" of the controller. DIRECT
 										                    //   means the output will increase when error is positive. REVERSE
@@ -57,15 +57,15 @@ class PID
 										  
 										  
   //Display functions ****************************************************************
-	double GetKp();						  // These functions query the pid for interal values.
-	double GetKi();						  //  they were created mainly for the pid front-end,
-	double GetKd();						  // where it's important to know what is actually 
+	float GetKp();						  // These functions query the pid for interal values.
+	float GetKi();						  //  they were created mainly for the pid front-end,
+	float GetKd();						  // where it's important to know what is actually 
 	// int GetMode();						  //  inside the PID.
 	int GetDirection();					//
   unsigned long windowStartTime;
 
-  double GetOutMax();
-  double GetOutMin();
+  float GetOutMax();
+  float GetOutMin();
 
   unsigned long GetSampleTime();
 
@@ -75,27 +75,27 @@ class PID
   private:
 	void Initialize();
 	
-	double dispKp;				        // * we'll hold on to the tuning parameters in user-entered 
-	double dispKi;				        //   format for display purposes
-	double dispKd;				        //
+	float dispKp;				        // * we'll hold on to the tuning parameters in user-entered 
+	float dispKi;				        //   format for display purposes
+	float dispKd;				        //
     
-	double kp;                   // * (P)roportional Tuning Parameter
-  double ki;                   // * (I)ntegral Tuning Parameter
-  double kd;                   // * (D)erivative Tuning Parameter
+	float kp;                   // * (P)roportional Tuning Parameter
+  float ki;                   // * (I)ntegral Tuning Parameter
+  float kd;                   // * (D)erivative Tuning Parameter
 
 	int controllerDirection;
   int pOn;                     // * Proportional on Error (default) OR Proportional on Measurement
 
-  // double *myInput;              // * Pointers to the Input, Output, and Setpoint variables
-  // double *myOutput;             //   This creates a hard link between the variables and the 
-  // double *mySetpoint;           //   PID, freeing the user from having to constantly tell us
+  // float *myInput;              // * Pointers to the Input, Output, and Setpoint variables
+  // float *myOutput;             //   This creates a hard link between the variables and the 
+  // float *mySetpoint;           //   PID, freeing the user from having to constantly tell us
                                   //   what these values are.  with pointers we'll just know.
 			  
 	unsigned long lastTime;
-	double outputSum, lastInput;
+	float outputSum, lastInput;
 
 	unsigned long SampleTime;         // * millis
-	double outMin, outMax;
+	float outMin, outMax;
 	bool pOnE;                        // * Proportional on Error (default) OR Proportional on Measurement
 };
 #endif
