+++ src/functional/thermo_controll.cpp
@@ -32,15 +32,15 @@ void myPIDsInit() {
 void initOnBrewPID() {
   //fast, aggressvie PID
   int dt = HEAT_BREW_TIME_INTERVAL;
-  double min = 0.f;
-  double max = MAX_BOILER_ON_PERCENTAGE;
+  float min = 0.f;
+  float max = MAX_BOILER_ON_PERCENTAGE;
 
   //Ziegler–Nichols method PI
   //Tu = 31s
-  double Kp = 30.0;
-  double Ki = 1.161290322580645;
-  // double Ki = (1.2 * Kp) / 31.0;
-  double Kd = 0.0;
+  float Kp = 30.f;
+  float Ki = 1.161290322580645f;
+  // float Ki = (1.2 * Kp) / 31.0;
+  float Kd = 0.0;
 
   PIDGroupSingleton::getOnBrewPID().SetTunings(Kp, Ki, Kd);
   PIDGroupSingleton::getOnBrewPID().SetOutputLimits(min, max);
@@ -50,15 +50,15 @@ void initOnBrewPID() {
 void initOffBrewPID() {
   //slow, balanced PID
   int dt1 = HEAT_TIME_INTERVAL;
-  double min1 = 0.0;
-  double max1 = MAX_BOILER_ON_PERCENTAGE;
+  float min1 = 0.0;
+  float max1 = MAX_BOILER_ON_PERCENTAGE;
 
   //Ziegler–Nichols method PI
   //Tu = 31s
-  double Kp1 = 4.5;
-  double Ki1 = 0.1741935483870968;
-  // double Ki1 = (1.2 * Kp1) / 31.0;
-  double Kd1 = 0.0;
+  float Kp1 = 4.5;
+  float Ki1 = 0.1741935483870968f;
+  // float Ki1 = (1.2 * Kp1) / 31.0;
+  float Kd1 = 0.0;
 
   PIDGroupSingleton::getOffBrewPID().SetTunings(Kp1, Ki1, Kd1);
   PIDGroupSingleton::getOffBrewPID().SetOutputLimits(min1, max1);
@@ -113,7 +113,7 @@ float computeThermoCompensateEnergyByInletWater(float compensateTemp, float targ
     
     // float correctionTemp = caculateCorrection(heatState, currentTemp, targetTemp);
     // float correctionTemp = 0.f;
-    double deltaHeat = (double)deltaWater * (double)compensateTemp * (double)WATER_TEMP_RISE_POWER;
+    float deltaHeat = deltaWater * compensateTemp * WATER_TEMP_RISE_POWER;
     heatState.lastThermoCompensateHeat = deltaHeat;
 
     // add balance
@@ -159,10 +159,10 @@ float caculateCorrection(HeatState& heatState, float currentTemp, float setPoint
 }
 
 float computeHeaterConsumedEnergyAndDoHeat(HeatState& heatState, float currentTemp, float setPoint, float upperLimit, float downLimit, int timeInterval) {
-  double energyConsumed = 0.0;
+  float energyConsumed = 0.f;
   bool isMyOperation = heatState.isBoilerOperatorTC;
-  u_int32_t lastHeaterWaveTime = heatState.lastBoilerStateTimestamp;
-  u_int32_t currentTime = micros();
+  uint32_t lastHeaterWaveTime = heatState.lastBoilerStateTimestamp;
+  uint32_t currentTime = micros();
   //timer overflow
   if (currentTime < lastHeaterWaveTime) {
     heatState.lastBoilerStateTimestamp = currentTime;
@@ -172,7 +172,7 @@ float computeHeaterConsumedEnergyAndDoHeat(HeatState& heatState, float currentTe
   if (deltaTime >= timeInterval * 1000UL) {
     if (isMyOperation && heatState.lastBoilerState) {
 
-      energyConsumed = HEATER_POWER * deltaTime * 0.000001;
+      energyConsumed = HEATER_POWER * deltaTime * 0.000001f;
       heatState.lastThermoHeaterConsumed = energyConsumed;
 
       // reduce balance
@@ -187,7 +187,7 @@ float computeHeaterConsumedEnergyAndDoHeat(HeatState& heatState, float currentTe
 
 void driveHeaterByEnergyBalance(HeatState& heatState, float currentTemp, float setPoint, float upperLimit, float downLimit) {
   bool boilerOperatorTC = true;
-  if (heatState.heatBalancePool <= 0.0) {
+  if (heatState.heatBalancePool <= 0.f) {
     if (currentTemp >= setPoint - downLimit) {
       turnOffBoiler(heatState, boilerOperatorTC);
     }
@@ -206,12 +206,12 @@ void driveHeaterByEnergyBalance(HeatState& heatState, float currentTemp, float s
 }
 
 // it's realtime job, cant't accumulate by time.
-double doPIDAdjust(float targetTemp, PID& pid, const SensorState& currentState, HeatState& heatState) {
+float doPIDAdjust(float targetTemp, PID& pid, const SensorState& currentState, HeatState& heatState) {
   
   //do pid
-  double outMin = pid.GetOutMin();
+  float outMin = pid.GetOutMin();
    //compute output
-  double output = pid.Compute(currentState.temperature, targetTemp);
+  float output = pid.Compute(currentState.temperature, targetTemp);
   //only if output is legal, update the heatState's output to new value
   if (output >= outMin) {
     heatState.lastPidOutputTimestamp = micros();
@@ -227,9 +227,9 @@ double doPIDAdjust(float targetTemp, PID& pid, const SensorState& currentState, 
   return output;
 }
 
-double doPIDAdjustWithLimit(float targetTemp, float downLimit, float upperLimit, PID& pidController, const SensorState& currentState, HeatState& heatState) {
+float doPIDAdjustWithLimit(float targetTemp, float downLimit, float upperLimit, PID& pidController, const SensorState& currentState, HeatState& heatState) {
   float currentTemp = currentState.temperature;
-  double out = pidController.GetOutMin() - 2.0;
+  float out = pidController.GetOutMin() - 2.f;
   if (currentTemp < targetTemp - downLimit) {
     turnOnBoiler(heatState);
   }
@@ -247,7 +247,7 @@ void pulseHeaters(HeatState& heatState, unsigned long powerAdjustCycle) {
   bool boilerOn = heatState.lastBoilerState;
   // output%, so * 0.01
   // powerAdjustCycle is in micro seconds
-  unsigned long onTime = round(heatState.pidOutput * 0.01 * powerAdjustCycle);
+  unsigned long onTime = round(heatState.pidOutput * 0.01f * powerAdjustCycle);
   uint32_t currentTime = micros();
   //timer overflow
   if (currentTime < lastHeaterWaveTime) {
